an idea of how to represent colorable edges
This is not good, though, because the color cannot be changed in-place in an
array, due to the way structs work.

(Actually, compared to my own priors, it's due to the way *arrays* work. In C#,
this works with arrays, but not with other generic collections which would
reutrn copies. In Crystal, Arrays are like other collections and indexing
returns a copy of the value.)