added unrelated Prim implementation, to then modify
I need to check what the actual correct result is for msf.in2 and
other test cases. For single-component test cases, any existing
solution for finding a minimum spanning tree, even if it does not
work for minimum spanning forests in general, is adequate.

As an initial check, it is sufficient to merely verify that the
total weight is correct, so an existing solution to that problem
that I believe is accurate should be helpful.

That problem is one I've solved a number of times with multiple
algorithms (though usually Kruskal's or Prim's) in a number of
different languages, for competitive programming practice
challenges on multiple sites.

This file this commit adds was an accepted HackerRank submission.

It requires modification to be usable for checking the weight of
minimum spanning trees in the format accepted by msf.cr:

(1) Currently it uses 1-based indexing but it needs to use 0-based
indexing.

(2) Currently the number of edges it reads is a parameter supplied
on the first line. So the input format must be changed so that this
value is not expected on the first line of input, and instead edges
are read from standard input until end-of-file is reached.