modified separate Prim implementation for this input format
As I expected but wanted to check, the Kruskal implementation in
msf.cr appears correct, while the recently added Prim
implementation in msf.cr appears wrong.

So I'll focus my next efforts on identifying the bug there.

It would be nicer if there were an actual debugger for Crystal
(maybe there is but I just don't know about it?) but I should be
able to make do. :)