removed msf stuff, it's in its own repo
The msf (minimum spanning forests) program, and related programs
and files, are now in their own repository, called msf:

    https://github.com/EliahKagan/msf

That repository has their history, but not the history of other
files in this repository (done by rebasing; see the msf readme).

This repository retains the history as well, but the files are
deleted as of this commit.